# .github/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hellion Control API
  version: 0.1.0
  description: |
    HTTP API for creating and monitoring security test runs.
    Workers consume jobs from NATS and write events back to Redis.

servers:
  - url: http://localhost:8080
    description: Local docker-compose stack

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /runs:
    get:
      summary: List runs
      operationId: listRuns
      responses:
        "200":
          description: List of runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"

    post:
      summary: Create run
      operationId: createRun
      description: |
        Creates a single run and queues one job per target on the NATS subject
        `hellion.jobs.http.{scope_id}`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "200":
          description: Run queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          description: Invalid request or unknown test pack

  /runs/bulk:
    post:
      summary: Create many runs
      operationId: createBulkRuns
      description: |
        Creates one run per target. Useful for benchmarks and load tests.
        When `targets` is omitted, `target` is repeated `count` times.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkRunRequest"
      responses:
        "200":
          description: Runs queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: queued
                  created:
                    type: integer
                    example: 100
        "400":
          description: Invalid request or unknown test pack

  /runs/{run_id}:
    get:
      summary: Get run status
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          description: Run not found

    delete:
      summary: Delete run
      operationId: deleteRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                  status:
                    type: string
                    example: deleted
        "404":
          description: Run not found

  /runs/{run_id}/events:
    get:
      summary: Get run events
      operationId: getRunEvents
      description: Returns the full event history for a run as newline-delimited JSON (NDJSON).
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: NDJSON event history
          content:
            application/x-ndjson:
              schema:
                type: string
        "404":
          description: Run not found

  /runs/{run_id}/cancel:
    post:
      summary: Cancel run
      operationId: cancelRun
      description: Marks a run as cancelled. Jobs already in flight may still complete.
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                  status:
                    type: string
                    example: cancelled
        "404":
          description: Run not found

  /admin/queue/clear:
    post:
      summary: Clear pending job queue
      operationId: clearQueue
      responses:
        "200":
          description: Queue cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: queue cleared

  /admin/runs/clear:
    post:
      summary: Clear all run history
      operationId: clearRuns
      responses:
        "200":
          description: Runs cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: runs cleared

components:
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        example: run_20260605175954

  schemas:
    CreateRunRequest:
      type: object
      required:
        - scope_id
        - targets
        - test_pack
      properties:
        scope_id:
          type: string
          description: Scope identifier; must match a worker's configured scope.
          example: local-juice-shop
        targets:
          type: array
          items:
            type: string
            format: uri
          example:
            - http://juice-shop:3000
        test_pack:
          type: string
          description: Test pack filename without `.yaml` extension.
          example: juice-shop-detect

    BulkRunRequest:
      type: object
      required:
        - scope_id
        - test_pack
      properties:
        scope_id:
          type: string
          example: local-juice-shop
        target:
          type: string
          format: uri
          description: Single target to repeat when `targets` is omitted.
          example: http://juice-shop:3000
        targets:
          type: array
          items:
            type: string
            format: uri
        test_pack:
          type: string
          example: juice-shop-detect
        count:
          type: integer
          minimum: 1
          default: 1
          description: Number of runs to create when using `target` instead of `targets`.

    Run:
      type: object
      properties:
        run_id:
          type: string
        status:
          type: string
          enum:
            - queued
            - running
            - completed
            - cancelled
            - failed
        outcome:
          type: string
          enum:
            - unknown
            - not_exploitable
            - potentially_exploitable
            - exploitable
            - error
        scope_id:
          type: string
        test_pack:
          type: string
        created_at:
          type: string
          format: date-time

    Event:
      type: object
      properties:
        event_type:
          type: string
          example: finding.created
        run_id:
          type: string
        target:
          type: string
          nullable: true
        message:
          type: string
        severity:
          type: string
          nullable: true